move request_and_validate from client.utils to openapi_tools (add back-ref)
